+++ packages/webapp-yakshaving/source/index.ts
@@ -2,9 +2,10 @@ import yargs from 'yargs-parser'
 
 import path from 'path'
 import module from 'module'
-import fs from 'fs'
 
-import { validate } from './validator'
+import { validate } from './utils/validator'
+import { developmentMode } from './modes/development'
+import { productionMode } from './modes/production'
 
 import type {
 	rollup as RollupFn,
@@ -12,7 +13,6 @@ import type {
 	RollupOptions as RollupInputOptions,
 	OutputOptions as RollupOutputOptions,
 	Plugin,
-	RollupWatcher,
 	PluginImpl,
 } from 'rollup'
 
@@ -202,311 +202,4 @@ export async function configure(options: ConfigurationOptions) {
 			rollup,
 		})
 	}
-}
-
-/**
- * Start development mode
- * @param options - Options
- * @param options.dependenciesBundleOptions - Options for dependency bundle generation
- * @param options.watchModeOptions - Options for watch mode
- */
-async function developmentMode({
-	dependenciesBundleOptions,
-	watchModeOptions,
-}: {
-	dependenciesBundleOptions: Parameters<typeof createDependenciesBundle>[0]
-	watchModeOptions: Parameters<typeof startWatchMode>[0]
-}) {
-	// Create dependencies bundle
-	await createDependenciesBundle(dependenciesBundleOptions)
-	// Start watch mode
-	await startWatchMode(watchModeOptions)
-}
-
-/**
- * Create bundle of dependencies
- * @param options - Options
- * @param options.dependencies - Dependencies field from user's manifest
- * @param options.userRequire - User's require method
- * @param options.plugins - Tuple of commonjs, node-resolve and replace plugin
- * @param options.outputDirectory - Output directory for dependencies
- * @param options.rollup - Rollup method
- */
-async function createDependenciesBundle({
-	dependencies,
-	userRequire,
-	plugins: [pluginCommonJs, pluginNodeResolve, pluginReplace],
-
-	outputDirectory,
-
-	rollup,
-}: {
-	dependencies: Record<string, string>
-	userRequire: NodeRequire
-	plugins: [
-		typeof PluginCommonJs,
-		typeof PluginNodeResolve,
-		typeof PluginReplace
-	]
-
-	outputDirectory: string
-
-	rollup: typeof RollupFn
-}) {
-	// Caching
-	const cacheFilePath = path.join(outputDirectory, 'cache.json')
-	const cache: string[] = fs.existsSync(cacheFilePath)
-		? require(cacheFilePath)
-		: // In case file doesn't exist create the output directory
-		  (await fs.promises.mkdir(outputDirectory, { recursive: true }), [])
-	const updatedCache: string[] = []
-
-	// Dependencies entry points
-	const entryPoints: Record<string, string> = {}
-
-	for (const [name, descriptor] of Object.entries(dependencies)) {
-		// If dependency isn't already cached
-		// Resolve its esm entry and add to entryPoints
-		if (!cache.includes(name + descriptor)) {
-			const root = (...args: string[]) => path.join(name, ...args)
-			const { main, module, type } = userRequire(root('package.json'))
-
-			const modulePath = root(type === 'module' ? main : module || main)
-			entryPoints[name] = userRequire.resolve(modulePath)
-		}
-
-		// Update Cache
-		updatedCache.push(name + descriptor)
-	}
-
-	// Write updated cache
-	await fs.promises.writeFile(cacheFilePath, JSON.stringify(updatedCache))
-
-	// If all dependencies are cached return
-	if (Object.keys(entryPoints).length === 0) return
-
-	// Rollup input options
-	const inputOptions: RollupInputOptions = {
-		input: entryPoints,
-		context: 'window',
-		plugins: [
-			pluginCommonJs(),
-			pluginNodeResolve({ preferBuiltins: false }),
-			pluginReplace({
-				values: {
-					'process.env.NODE_ENV': '"development"',
-				},
-			}),
-		],
-	}
-
-	// Rollup output options
-	const outputOptions: RollupOutputOptions = {
-		dir: outputDirectory,
-		format: 'es',
-		entryFileNames: '[name].js',
-		sourcemap: true,
-		exports: 'named',
-	}
-
-	// Generate bundle
-	const bundle = await rollup(inputOptions)
-	await bundle.write(outputOptions)
-}
-
-/**
- * Start watch mode for application
- * @param options - Options
- * @param options.entryPoint - Entry point for application
- * @param options.dependencyMap - Map of dependency and their relative location
- * @param options.requiredPlugins - Object of required plugins and their configuration
- * @param options.extraPlugins - Additional plugins
- * @param options.outputDirectory - Output directory for bundle
- * @param options.watch - Rollup watch method
- * @param options.additionalRollupSettings - Additional rollup settings
- */
-async function startWatchMode({
-	entryPoint,
-	dependencyMap,
-
-	requiredPlugins: {
-		commonJs: [pluginCommonJs, commonJsUserOpts],
-		nodeResolve: [pluginNodeResolve, nodeResolveUserOpts],
-		sucrase: [pluginSucrase, sucraseUserOpts],
-		typescript: [pluginTypescript, typescriptUserOpts],
-	},
-	extraPlugins,
-
-	outputDirectory,
-	watch,
-
-	additionalRollupSettings,
-}: {
-	entryPoint: string
-	dependencyMap: Record<string, string>
-
-	requiredPlugins: {
-		commonJs: [typeof PluginCommonJs, Parameters<typeof PluginCommonJs>[0]?]
-		nodeResolve: [
-			typeof PluginNodeResolve,
-			Parameters<typeof PluginNodeResolve>[0]?
-		]
-		sucrase: [typeof PluginSucrase, Parameters<typeof PluginSucrase>[0]?]
-		typescript: [
-			typeof PluginTypescript,
-			Parameters<typeof PluginTypescript>[0]?
-		]
-	}
-	extraPlugins: Plugin[]
-
-	outputDirectory: string
-	watch: typeof WatchFn
-
-	additionalRollupSettings?: {
-		input?: RollupInputOptions
-		output?: RollupOutputOptions
-	}
-}) {
-	// Input options
-	const inputOptions: RollupInputOptions = {
-		input: entryPoint,
-		context: 'window',
-		plugins: [
-			pluginCommonJs(commonJsUserOpts),
-			pluginNodeResolve({
-				extensions: ['.mjs', '.js', '.json', '.node', '.ts', '.tsx'],
-				preferBuiltins: false,
-				...nodeResolveUserOpts,
-			}),
-
-			...extraPlugins,
-
-			pluginSucrase({
-				transforms: ['typescript', 'jsx'],
-				...sucraseUserOpts,
-			}),
-			pluginTypescript({
-				noEmit: true,
-				...typescriptUserOpts,
-			} as Parameters<typeof PluginTypescript>[0]),
-		],
-		external: Object.keys(dependencyMap),
-		...additionalRollupSettings?.input,
-	}
-
-	// Output options
-	const outputOptions: RollupOutputOptions = {
-		dir: outputDirectory,
-		format: 'es',
-		entryFileNames: '[name].bundle.js',
-		sourcemap: true,
-		paths: dependencyMap,
-		...additionalRollupSettings?.output,
-	}
-
-	const watchOptions = { clearScreen: false }
-
-	// Watcher
-	const watcher = watch({
-		...inputOptions,
-		output: [outputOptions],
-		watch: watchOptions,
-	})
-
-	// Listener for watcher :P
-	listenTheWatcher(watcher)
-}
-
-function listenTheWatcher(watcher: RollupWatcher) {
-	watcher.on('event', (event) => {
-		console.log(event)
-	})
-}
-
-/**
- * Start production mode
- * @param options - Options
- * @param options.entryPoint - Entry point for application
- * @param options.requiredPlugins - Object of required plugins and their configuration
- * @param options.extraPlugins - Additional Plugins
- * @param options.outputDirectory - Output directory for bundle
- * @param options.rollup - Rollup method
- * @param options.additionalRollupSettings - Additional rollup settings
- */
-async function productionMode({
-	entryPoint,
-
-	requiredPlugins: {
-		commonJs: [pluginCommonJs, commonJsUserOpts],
-		nodeResolve: [pluginNodeResolve, nodeResolveUserOpts],
-		typescript: [pluginTypescript, typescriptUserOpts],
-		terser: [pluginTerser, terserUserOpts],
-		replace: [pluginReplace],
-	},
-	extraPlugins,
-
-	outputDirectory,
-	rollup,
-
-	additionalRollupSettings,
-}: {
-	entryPoint: string
-
-	requiredPlugins: {
-		commonJs: [typeof PluginCommonJs, Parameters<typeof PluginCommonJs>[0]?]
-		nodeResolve: [
-			typeof PluginNodeResolve,
-			Parameters<typeof PluginNodeResolve>[0]?
-		]
-		typescript: [
-			typeof PluginTypescript,
-			Parameters<typeof PluginTypescript>[0]?
-		]
-		terser: [typeof PluginTerser, Parameters<typeof PluginTerser>[0]?]
-		replace: [typeof PluginReplace, Parameters<typeof PluginReplace>[0]?]
-	}
-	extraPlugins: Plugin[]
-
-	outputDirectory: string
-	rollup: typeof RollupFn
-
-	additionalRollupSettings?: {
-		input?: RollupInputOptions
-		output?: RollupOutputOptions
-	}
-}) {
-	const plugins = [
-		pluginCommonJs(commonJsUserOpts),
-		pluginNodeResolve({
-			extensions: ['.mjs', '.js', '.json', '.node', '.ts', '.tsx'],
-			preferBuiltins: false,
-			...nodeResolveUserOpts,
-		}),
-		pluginTypescript(typescriptUserOpts),
-		pluginReplace({
-			values: {
-				'process.env.NODE_ENV': '"production"',
-			},
-		}),
-		pluginTerser(terserUserOpts),
-		...extraPlugins,
-	]
-
-	const inputOptions: RollupInputOptions = {
-		input: entryPoint,
-		context: 'window',
-		plugins,
-		...additionalRollupSettings?.input,
-	}
-
-	const outputOptions: RollupOutputOptions = {
-		dir: outputDirectory,
-		format: 'es',
-		entryFileNames: '[name].bundle.js',
-		sourcemap: true,
-		...additionalRollupSettings?.output,
-	}
-
-	const bundle = await rollup(inputOptions)
-	await bundle.write(outputOptions)
 }
